requirements.txt
asgiref==3.5.2
Brotli==1.0.9
dj-database-url==1.2.0
Django==4.1.4
djangorestframework==3.14.0
gunicorn==20.1.0
mysqlclient==2.1.1
psycopg2-binary==2.9.5
PyMySQL==1.0.2
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.2.0
